moved decorators to separate file, and correctly marked async/non-async views
